# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Attachment": {
        "description": "Attachment with filename and URL",
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "url"
        ],
        "title": "Attachment",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NoteChainRead": {
        "description": "Read response model",
        "properties": {
          "note_chain_id": {
            "format": "uuid",
            "title": "Note Chain Id",
            "type": "string"
          },
          "read_permission": {
            "$ref": "#/components/schemas/NotePermission",
            "default": "Admin"
          },
          "write_permission": {
            "$ref": "#/components/schemas/NotePermission",
            "default": "Admin"
          }
        },
        "required": [
          "note_chain_id"
        ],
        "title": "NoteChainRead",
        "type": "object"
      },
      "NoteCreate": {
        "description": "Create request model - chain_id and user_id set by the service",
        "properties": {
          "attachments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Attachment"
            },
            "title": "Attachments",
            "type": "array"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "NoteCreate",
        "type": "object"
      },
...
      "NotePermission": {
        "description": "Controls who can read/write on a note chain",
        "enum": [
          "Admin",
          "All"
        ],
        "title": "NotePermission",
        "type": "string"
      },
      "NoteRead": {
        "description": "Read response model",
        "properties": {
          "attachments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Attachment"
            },
            "title": "Attachments",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "is_system": {
            "default": false,
            "title": "Is System",
            "type": "boolean"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "note_chain_id": {
            "format": "uuid",
            "title": "Note Chain Id",
            "type": "string"
          },
          "note_id": {
            "format": "uuid",
            "title": "Note Id",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "note_chain_id",
          "message",
          "note_id"
        ],
        "title": "NoteRead",
        "type": "object"
      },
      "NoteUpdate": {
        "description": "Update request model - only message can be edited",
        "properties": {
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "NoteUpdate",
        "type": "object"
      },
...
      "RouteDetailRead": {
        "description": "GET /routes/{route_id} response: the bare route plus its ordered stops.\n\nStops are assembled with snapshot-over-live precedence. See\nRouteStopDetailRead.",
        "properties": {
          "cloned_from_route_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloned From Route Id"
          },
          "driver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "encoded_polyline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encoded Polyline"
          },
          "ends_at_warehouse": {
            "default": false,
            "title": "Ends At Warehouse",
            "type": "boolean"
          },
          "length": {
            "minimum": 0,
            "title": "Length",
            "type": "number"
          },
          "name": {
            "default": "",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "note_chain_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Chain Id"
          },
          "notes": {
            "default": "",
            "maxLength": 1000,
            "title": "Notes",
            "type": "string"
          },
          "polyline_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polyline Updated At"
          },
          "route_group_id": {
            "format": "uuid",
            "title": "Route Group Id",
            "type": "string"
          },
          "route_id": {
            "format": "uuid",
            "title": "Route Id",
            "type": "string"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/RouteStopDetailRead"
            },
            "title": "Stops",
            "type": "array"
          }
        },
        "required": [
          "length",
          "route_group_id",
          "route_id"
        ],
        "title": "RouteDetailRead",
        "type": "object"
      },
...
      "RoutePatchRequest": {
        "description": "Request body for PATCH /routes/{route_id}.\n\nAll fields are optional - only provided fields will be updated.\nIf location_ids is provided, the route stops will be fully replaced\nand the routing algorithm will be re-run to update polyline + mileage.",
        "properties": {
          "driver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "location_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Ids"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          }
        },
        "title": "RoutePatchRequest",
        "type": "object"
      },
      "RouteRead": {
        "description": "Read response model",
        "properties": {
          "cloned_from_route_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloned From Route Id"
          },
          "driver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "encoded_polyline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encoded Polyline"
          },
          "ends_at_warehouse": {
            "default": false,
            "title": "Ends At Warehouse",
            "type": "boolean"
          },
          "length": {
            "minimum": 0,
            "title": "Length",
            "type": "number"
          },
          "name": {
            "default": "",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "note_chain_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Chain Id"
          },
          "notes": {
            "default": "",
            "maxLength": 1000,
            "title": "Notes",
            "type": "string"
          },
          "polyline_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polyline Updated At"
          },
          "route_group_id": {
            "format": "uuid",
            "title": "Route Group Id",
            "type": "string"
          },
          "route_id": {
            "format": "uuid",
            "title": "Route Id",
            "type": "string"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          }
        },
        "required": [
          "length",
          "route_group_id",
          "route_id"
        ],
        "title": "RouteRead",
        "type": "object"
      },
...
      "RouteStopDetailRead": {
        "description": "A single ordered stop on a route, assembled with snapshot-over-live\nfield precedence: for a frozen (past) route the route_stop_snapshot wins,\notherwise the live Location is read. Embedded in RouteDetailRead.\n\nNotes are intentionally NOT embedded: note_chain_id points at the stop's\nnote chain, fetched separately via GET /note-chains/{id}/notes.",
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "boxes": {
            "title": "Boxes",
            "type": "integer"
          },
          "contact_name": {
            "title": "Contact Name",
            "type": "string"
          },
          "note_chain_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Chain Id"
          },
          "phone_primary": {
            "title": "Phone Primary",
            "type": "string"
          },
          "phone_secondary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Secondary"
          },
          "stop_number": {
            "title": "Stop Number",
            "type": "integer"
          }
        },
        "required": [
          "stop_number",
          "address",
          "contact_name",
          "phone_primary",
          "boxes"
        ],
        "title": "RouteStopDetailRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the Food4Kids application",
    "title": "Food4Kids API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/note-chains/{note_chain_id}": {
      "delete": {
        "description": "Delete a note chain and all its notes (admin only)",
        "operationId": "delete_note_chain",
        "parameters": [
          {
            "in": "path",
            "name": "note_chain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Chain Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Note Chain",
        "tags": [
          "note-chains"
        ]
      },
      "get": {
        "description": "Get a note chain by ID (checks read permission)",
        "operationId": "get_note_chain",
        "parameters": [
          {
            "in": "path",
            "name": "note_chain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Chain Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteChainRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Note Chain",
        "tags": [
          "note-chains"
        ]
      }
    },
    "/note-chains/{note_chain_id}/notes": {
      "get": {
        "description": "Get notes for a chain with pagination.",
        "operationId": "get_notes",
        "parameters": [
          {
            "in": "path",
            "name": "note_chain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Chain Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NoteRead"
                  },
                  "title": "Response Get Notes",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Notes",
        "tags": [
          "note-chains"
        ]
      },
      "post": {
        "description": "Add a note to a chain",
        "operationId": "create_note",
        "parameters": [
          {
            "in": "path",
            "name": "note_chain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Chain Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Note",
        "tags": [
          "note-chains"
        ]
      }
    },
    "/note-chains/{note_chain_id}/notes/{note_id}": {
      "delete": {
        "description": "Delete a note (author or admin only)",
        "operationId": "delete_note",
        "parameters": [
          {
            "in": "path",
            "name": "note_chain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Chain Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Note",
        "tags": [
          "note-chains"
        ]
      },
      "patch": {
        "description": "Edit a note's message (author or admin only)",
        "operationId": "update_note",
        "parameters": [
          {
            "in": "path",
            "name": "note_chain_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Chain Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Note",
        "tags": [
          "note-chains"
        ]
      }
    },
...
    "/routes/{route_id}": {
      "delete": {
        "description": "Delete a route by its unique identifier.\n\nParameters:\n    route_id (UUID): The unique identifier of the route to delete.\n    session (AsyncSession): The database session dependency.\n    _ (bool): Indicates that the user is authenticated as a driver (injected by dependency).\n\nReturns:\n    None. Responds with HTTP 204 No Content on successful deletion.",
        "operationId": "delete_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Route",
        "tags": [
          "routes"
        ]
      },
      "get": {
        "description": "Get a route by its unique identifier, with its ordered stops embedded.\n\nEach stop carries its sequence #, address, contact name, phone (+ secondary),\nbox count, and a note_chain_id reference. Stops are sourced live from\nLocation for upcoming routes and from frozen route_stop_snapshots for past\nroutes. Notes are not embedded: fetch them via GET /note-chains/{id}/notes.\n\nParameters:\n    route_id (UUID): The unique identifier of the route to GET.\n    session (AsyncSession): The database session dependency.\n\nReturns:\n    The route with its ordered stops. Responds with HTTP 200 OK on success.",
        "operationId": "get_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Route",
        "tags": [
          "routes"
        ]
      },
      "patch": {
        "description": "Update a route's metadata (name, notes) and/or its stop order/locations.\n\nIf location_ids is provided in the request body:\n- The route's stops are fully replaced with the new ordered list.\n- The routing algorithm is re-run to compute the new polyline and mileage.\n- This will affect ALL route groups that share this route.\n\nIf only name/notes are provided, stops and mileage are left unchanged.",
        "operationId": "update_route",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Route",
        "tags": [
          "routes"
        ]
      }
    },
...
  }
}